# Fix the issue when calculating λ₂ by using Hermitian matrix