@weymouth 2024-04-20
I don't see any issue here, so I'll merge. Out of curiosity, what tipped you off to this issue? What did the field look like? Can we get a before and after picture or something?