@TzuYaoHuang 2024-04-20
It's more like a platform-dependent bug. When I run TGV on my Mac, it will report a bug like the following line but after making the change, everything is fine.

```
ERROR: LoadError: TaskFailedException

    nested task error: Only hermitian matrices are diagonalizable by *StaticArrays*. Non-Hermitian matrices should be converted to `Array` first.
    Stacktrace:
     [1] error(s::String)
       @ Base ./error.jl:35
     [2] #eigvals#515
       @ ~/.julia/packages/StaticArrays/EHHaF/src/eigen.jl:12 [inlined]
     [3] eigvals
       @ ~/.julia/packages/StaticArrays/EHHaF/src/eigen.jl:6 [inlined]
     [4] λ₂(I::CartesianIndex{3}, u::Array{Float32, 4})
       @ WaterLily ~/.julia/packages/WaterLily/CAGCc/src/Metrics.jl:43
     [5] macro expansion
       @ ~/.julia/packages/WaterLily/CAGCc/src/util.jl:99 [inlined]
     [6] cpu_
       @ ~/.julia/packages/KernelAbstractions/zPAn3/src/macros.jl:287 [inlined]
     [7] (::var"#cpu_##kern#225#3")(__ctx__::KernelAbstractions.CompilerMetadata{KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.NoDynamicCheck, CartesianIndex{3}, CartesianIndices{3, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}, Base.OneTo{Int64}}}, KernelAbstractions.NDIteration.NDRange{3, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.StaticSize{(64, 1, 1)}, CartesianIndices{3, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}, Base.OneTo{Int64}}}, Nothing}}, a::Array{Float32, 3}, u::Array{Float32, 4}, L::Int64, U::Int64, I0::CartesianIndex{3})
       @ Main ./none:0
     [8] __thread_run(tid::Int64, len::Int64, rem::Int64, obj::KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.StaticSize{(64,)}, KernelAbstractions.NDIteration.DynamicSize, var"#cpu_##kern#225#3"}, ndrange::Tuple{Int64, Int64, Int64}, iterspace::KernelAbstractions.NDIteration.NDRange{3, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.StaticSize{(64, 1, 1)}, CartesianIndices{3, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}, Base.OneTo{Int64}}}, Nothing}, args::Tuple{Array{Float32, 3}, Array{Float32, 4}, Int64, Int64, CartesianIndex{3}}, dynamic::KernelAbstractions.NDIteration.NoDynamicCheck)
       @ KernelAbstractions ~/.julia/packages/KernelAbstractions/zPAn3/src/cpu.jl:115
     [9] (::KernelAbstractions.var"#18#21"{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.StaticSize{(64,)}, KernelAbstractions.NDIteration.DynamicSize, var"#cpu_##kern#225#3"}, Tuple{Int64, Int64, Int64}, KernelAbstractions.NDIteration.NDRange{3, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.StaticSize{(64, 1, 1)}, CartesianIndices{3, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}, Base.OneTo{Int64}}}, Nothing}, Tuple{Array{Float32, 3}, Array{Float32, 4}, Int64, Int64, CartesianIndex{3}}, KernelAbstractions.NDIteration.NoDynamicCheck, Int64})()
       @ KernelAbstractions ~/.julia/packages/KernelAbstractions/zPAn3/src/cpu.jl:90

...and 3 more exceptions.

Stacktrace:
 [1] sync_end(c::Channel{Any})
   @ Base ./task.jl:448
 [2] macro expansion
   @ ./task.jl:480 [inlined]
 [3] __run(obj::KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.StaticSize{(64,)}, KernelAbstractions.NDIteration.DynamicSize, var"#cpu_##kern#225#3"}, ndrange::Tuple{Int64, Int64, Int64}, iterspace::KernelAbstractions.NDIteration.NDRange{3, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.StaticSize{(64, 1, 1)}, CartesianIndices{3, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}, Base.OneTo{Int64}}}, Nothing}, args::Tuple{Array{Float32, 3}, Array{Float32, 4}, Int64, Int64, CartesianIndex{3}}, dynamic::KernelAbstractions.NDIteration.NoDynamicCheck, static_threads::Bool)
   @ KernelAbstractions ~/.julia/packages/KernelAbstractions/zPAn3/src/cpu.jl:89
 [4] (::KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.StaticSize{(64,)}, KernelAbstractions.NDIteration.DynamicSize, var"#cpu_##kern#225#3"})(::Array{Float32, 3}, ::Vararg{Any}; ndrange::Tuple{Int64, Int64, Int64}, workgroupsize::Nothing)
   @ KernelAbstractions ~/.julia/packages/KernelAbstractions/zPAn3/src/cpu.jl:44
 [5] macro expansion
   @ ~/.julia/packages/WaterLily/CAGCc/src/util.jl:102 [inlined]
 [6] λ₂!(dat::Array{Float32, 3}, sim::Simulation)
   @ Main ~/Documents/Github/WaterLily.jl/examples/ThreeD_TaylorGreenVortex.jl:25
 [7] makie_video!(makie_plot::var"#12#13", sim::Simulation, dat::Array{Float32, 3}, obs_update!::typeof(λ₂!); remeasure::Bool, name::String, duration::Int64, step::Float64, framerate::Int64, compression::Int64)
   @ Main ~/Documents/Github/WaterLily.jl/examples/ThreeD_Plots.jl:5
 [8] macro expansion
   @ ./timing.jl:279 [inlined]
 [9] top-level scope
   @ ~/Documents/Github/WaterLily.jl/examples/ThreeD_TaylorGreenVortex.jl:269
in expression starting at /Users/tzuyaohuang/Documents/Github/WaterLily.jl/examples/ThreeD_TaylorGreenVortex.jl:29
```